Create command line parameter to force disabled tests to run Create command line parameter to control mouse animation Improve message shown when running a test case
